Fix all compiler warnings produced by gcc.
This allows the linux and NaCl builds (and perhaps others)
to be compiles with -Wall and -Werror, and it makes the build
output much less noisy.

+++ cocos2dx/tilemap_parallax_nodes/CCTMXXMLParser.cpp
@@ -517,7 +517,7 @@ void CCTMXMapInfo::startElement(void *ctx, const char *name, const char **atts)
         // Parse everything automatically
         const char* pArray[] = {"name", "type", "width", "height", "gid"};
         
-        for( int i = 0; i < sizeof(pArray)/sizeof(pArray[0]); ++i )
+        for(size_t i = 0; i < sizeof(pArray)/sizeof(pArray[0]); ++i )
         {
             const char* key = pArray[i];
             CCString* obj = new CCString(valueForKey(key, attributeDict));
